rename activeBoardReducer to activeBoardsReducer
