+++ src/state/activeBoards/activeBoardsReducer.test.ts
@@ -1,12 +1,10 @@
-import activeBoardReducer from './activeBoardsReducer';
-import { setAllowedBoards } from './activeBoardActions';
-
-// TODO: rename activeBoardReducer file: remove s.
+import activeBoardsReducer from './activeBoardsReducer';
+import { setAllowedBoards } from './activeBoardsActions';
 
 describe( 'activeBoardsReducer', () => {
 
     it( 'should return init state', () => {
-        const initState = activeBoardReducer( undefined, {type: 'init'} );
+        const initState = activeBoardsReducer( undefined, {type: 'init'} );
         expect( initState ).not.toBeNull();
         expect( initState ).not.toBeUndefined();
     } );
@@ -14,7 +12,7 @@ describe( 'activeBoardsReducer', () => {
     it('should set the activeBoards to the points in the point array', () => {
         const newAllowedBoards = [{x: 0, y: 0}, {x: 2, y: 0}, {x: 2, y: 2}];
         const action = setAllowedBoards(newAllowedBoards);
-        const newState = activeBoardReducer(undefined, action);
+        const newState = activeBoardsReducer( undefined, action);
 
         expect(newState).toEqual(newAllowedBoards);
     });
