+++ src/state/configureStore.ts
@@ -6,7 +6,7 @@ import { composeWithDevTools } from 'redux-devtools-extension';
 import gameReducer from './game/gameReducer';
 import moveReducer from './moves/moveReducer';
 import { enableBatching } from 'redux-batched-actions';
-import activeBoardReducer from './activeBoards/activeBoardsReducer';
+import activeBoardsReducer from './activeBoards/activeBoardsReducer';
 import createSagaMiddleware from 'redux-saga';
 import { all, fork } from 'redux-saga/effects';
 import playerMovedSaga from './sagas/moveValidationSaga';
@@ -19,7 +19,7 @@ const rootreducer = combineReducers<AppState>(
         game: gameReducer,
         board: boardReducer,
         moves: moveReducer,
-        activeBoards: activeBoardReducer
+        activeBoards: activeBoardsReducer
     } );
 
 export function configureStore() {
