+++ src/state/sagas/checkGameFinishedSaga.ts
@@ -3,7 +3,7 @@ import { GenericAction } from '../AppState';
 import { CHECK_GAME_FINISHED, gameFinished } from '../game/gameAction';
 import { getBoards } from '../selectors/AppStateSelectors';
 import { getWinResult } from '../../util/CheckBoard';
-import { setAllowedBoards } from '../activeBoards/activeBoardActions';
+import { setAllowedBoards } from '../activeBoards/activeBoardsActions';
 
 function* checkIfGameFinished( action: GenericAction ) {
     const boards = yield select( getBoards );
