+++ src/state/sagas/test/activeBoardCalculationSaga.test.ts
@@ -1,7 +1,7 @@
 import { expectSaga } from 'redux-saga-test-plan';
 import { select } from 'redux-saga/effects';
 import { getBoards } from '../../selectors/AppStateSelectors';
-import { CALCULATE_ALLOWED_BOARDS, SET_ALLOWED_BOARDS } from '../../activeBoards/activeBoardActions';
+import { CALCULATE_ALLOWED_BOARDS, SET_ALLOWED_BOARDS } from '../../activeBoards/activeBoardsActions';
 import activeBoardsCalculationSaga from '../activeBoardsCalculationSaga';
 import unfinishedBoardMock from './unfinishedBoardMock';
 import { circleFinishedBoardMock } from './finishedBoardMock';
